Add item edit/create flow and timeline cursor marker
- EditItem window: full end-to-end implementation (LOD-aware date inputs,
  tags, characters, stories, books/chapters, save via bridge)
- New DB models and repos: BookItem, ChapterItem, BookRepo; extended
  ItemRepo with GetItemById, GetItemTags, GetItemCharacterAppearances,
  GetItemStoryRefs, GetItemChapterRefs, SaveItemFull
- Bridge: 7 new message handlers (GetItemForEdit, SaveItem, SearchTags,
  GetTimelineCharacters, GetTimelineStories, SearchBooks, GetBookChapters)
- Fixed WebView2 cache-folder conflict (separate subfolders per window)
  and double WebMessageReceived subscription in f_Timeline
- Replaced ShowDialog with Show for f_AddEditItem to avoid COM re-entrancy;
  WindowCloseRequested closes the form when Vue calls window.close()
- Canvas: itemClick and addItem emits wired through to OpenAddEditItemWindow
- Canvas: snap-to-tick cursor marker line on hover with LOD-aware date label;
  shift key overrides snapping; label and context menu time update on pan/zoom
- vite.config: added editItem entry point; fixed duplicate settings key
- models.ts: added CharacterItem, Tag, Story, Book, Chapter, ItemForEdit etc.

+++ StoryTimelineMk2/Database/TimelineItem.cs
@@ -16,8 +16,8 @@ namespace StoryTimelineMk2.Database
         public int Subtick { get; set; }
         public int OriginalSubtick { get; set; }
         public int EndYear { get; set; }
-        public int AbsoluteStart { get; set; }
-        public int AbsoluteEnd { get; set; }
+        public double AbsoluteStart { get; set; }
+        public double AbsoluteEnd { get; set; }
         public int EndSubtick { get; set; }
         public int OriginalEndSubtick { get; set; }
         public string BookTitle { get; set; }
